WAL improvements
- Use FileChannel to write WALs
- Refactor decoding/encoding to module/interface
- Configuration to enable WAL (disabled by default)
- Place WAL files under .wal directory to avoid polluting the repo
- Lazily create WAL file when notebook is changed
- Add timestamps to WAL messages
