+++ polynote-server/src/main/scala/polynote/RecoverLog.scala
@@ -1,22 +1,30 @@
 package polynote
 
+import java.nio.channels.FileChannel
 import java.nio.file.{Files, Paths, StandardOpenOption}
+import java.time.Instant
 
-import polynote.app.MainArgs
+import cats.effect.Effect
+import polynote.app.{Args, MainArgs}
 import polynote.kernel.logging.Logging
 import polynote.messages.{Message, Notebook, NotebookUpdate, ShortList}
 import polynote.server.AppEnv
 import zio.{Ref, Runtime, Task, UIO, ZIO}
 import zio.ZIO.effectTotal
 import zio.blocking.effectBlocking
-import zio.interop.catz._
 import fs2.Stream
-import polynote.server.repository.NotebookContent
+import polynote.server.repository.{FileBasedRepository, NotebookContent}
 import polynote.server.repository.format.ipynb.IPythonFormat
+import polynote.server.repository.fs.WAL
+import polynote.server.taskConcurrent
+import scodec.bits.ByteVector
+import scodec.stream.decode
+import scodec.codecs
+import scodec.stream.decode.StreamDecoder
 
 object RecoverLog {
 
-  def replay(messages: Stream[Task, Message], ref: Ref[Notebook], log: Logging.Service): UIO[Unit] = messages.evalMap {
+  def replay(messages: Stream[Task, (Instant, Message)], ref: Ref[Notebook], log: Logging.Service): UIO[Unit] = messages.map(_._2).evalMap {
     case nb: Notebook => ref.set(nb)
     case upd: NotebookUpdate => ref.update {
       nb => try {
@@ -33,13 +41,13 @@ object RecoverLog {
       log.error(Some("Error occurred while replaying the log; printing the final state anyway."), err)
   }
 
-  def main(implicit runtime: Runtime[Any]): ZIO[AppEnv, String, Int] = for {
-    args     <- ZIO.access[MainArgs](_.get.rest)
+  def main(implicit ev: Effect[Task]): ZIO[AppEnv, String, Int] = for {
+    args     <- ZIO.access[MainArgs](_.get[Args].rest)
     path     <- ZIO(args.head).flatMap(pathStr => effectBlocking(Paths.get(pathStr).toRealPath())).orDie
-    is       <- effectBlocking(Files.newInputStream(path, StandardOpenOption.READ)).orDie
+    is       <- effectBlocking(FileChannel.open(path, StandardOpenOption.READ)).orDie
     log      <- Logging.access
     _        <- Logging.info(s"Reading log entries from ${path}...")
-    messages  = scodec.stream.decode.many[Message](Message.codec).decodeInputStream[Task](is)
+    messages  = WAL.decoder.decodeMmap(is)
     ref      <- Ref.make(Notebook("", ShortList.Nil, None))
     _        <- replay(messages, ref, log)
     format    = new IPythonFormat
