+++ polynote-server/src/main/scala/polynote/server/repository/fs/NotebookFilesystem.scala
@@ -1,10 +1,15 @@
-package polynote.server.repository.fs
+package polynote.server
+package repository.fs
 
 import java.io.OutputStream
+import java.nio.ByteBuffer
 import java.nio.file.Path
 
 import polynote.kernel.BaseEnv
-import zio.RIO
+import polynote.messages.Message
+import scodec.bits.BitVector
+import zio.blocking.Blocking
+import zio.{RIO, Task, ZIO}
 
 /**
   * Encapsulates useful Notebook-related FS stuff.
@@ -17,7 +22,7 @@ trait NotebookFilesystem {
 
   def writeStringToPath(path: Path, content: String): RIO[BaseEnv, Unit]
 
-  def createLog(path: Path): RIO[BaseEnv, OutputStream]
+  def createLog(path: Path): RIO[BaseEnv, WAL.WALWriter]
 
   def list(path: Path): RIO[BaseEnv, List[Path]]
